requirements.txt
amqp==2.4.2
baidu-aip==2.2.13.0
billiard==3.5.0.5
celery==4.2.2
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-cors-headers==3.1.0
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.2
dramatiq==1.5.0
idna==2.7
kombu==4.3.0
libmagic==1.0
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
prometheus-client==0.2.0
python-magic==0.4.15
pytz==2018.5
requests==2.19.1
urllib3==1.23
vine==1.3.0
